requirements.txt
python-libtorrent
tkinter
ttkbootstrap
